requirements.txt
# JARVIS dependencies
# Install: pip install -r requirements.txt

sounddevice>=0.4.6
numpy>=1.24.0
SpeechRecognition>=3.10.0
edge-tts>=6.1.9
playsound==1.2.2
python-dotenv>=1.0.0